WIP: frequency and clutter
Main changes were to add a "radar files" key to the output of the
clutter frequency map which gives the user direct access to which files
were used in calculating the clutter frequency map. I have also added
the map_json_list function to clutter_frequency, which should be the
prefered method since the JSON file can define any time contiguous
or non-contiguous list of radar files to use in computing the clutter
frequency map.