WIP: specify sweeps
The user can now specify the minimum and maximum sweeps to use in
calculating the texture fields. This is being introduced because it may
in fact be better to only use the first sweep when computing the texture
fields, at least as far as the underlying histograms are concerned,
since insects at SGP act to broaden certain signals, and in fact create
a bimodal distribution in differential phase. That being said, insects
are often considered as clutter, so more testing may need to be done...