NEW: correct module
The correct module will include functions to fill in (by means of
interpolation) missing gates. As an example, we should necessarily be
filling in missing gates where the surrounding 8 gates all have
non-missing data. The user will have control over the size of the 2-D
window (default should be 3 x 3), as well as the minimum sample size
necessary for interpolating to the missing gate. More to come!