WIP: include PDFs in texture output
One important update is the inclusion of the actual probability density
function (PDF) in the output of histogram_from_json in
texture_fields.py. Note that a PDF, by definition, is such that when
integrated over the whole interval equals 1. The PDF instead of the
histogram or normalized histogram is the fundamental input to the
Bayesian method outlined in Rico-Ramirez and Cluckie (2008).

Rico-Ramirez, M. A. and I. D. Cluckie, 2008, "Classification of Ground
Clutter and Anomalous Propagation Using Dual-Polarization Weather
Radar".