WIP: radar gate interpolation
Some minor fixes to both the Fortran routine and its rebuild script, as
well as to the Python routine (basic_fixes.py) that calls the Fortran
routine. I quickly learned that if I declare an array in Fortran as
intent(inout) then I have to be careful to adhere to the data type or
else I'm going to get a ValueError.