WIP: Bayesian classifier
Like I said in a previous commit, my Bayesian classifier will likely
still need some work, but I have made great strides so far in getting
this operational. One major change is in removing the "rule" kwarg,
which didn't make sense for the Bayesian classifier as it is strictly a
product. More to come...