WIP: Bayes classifier
The user now has the ability to input only the moments fields in order
to classify radar gates (if they so choose). Furthermore, the user can
specify the minimum number of inputs necessary to make a "reliable"
classification. When the number of inputs is less than this number, the
gate is classified as missing. Finally, the user can specifiy any number
of inputs that (s)he would like to ignore (exclude) when classifying.
The Bayes classifier still does not incorporate the clutter frequency
map, so more work needs to be done...