WIP: gate filter input
The user is now able to input a gate filter during classification such
that filtered gates are not classified, i.e. set to missing. I am also
addressing a bug in which masked gates (pixels) were not actually
getting set to missing, and at the same time I am now much more explicit
with how I'm setting the classification array, so this should help with
readability...