NEW: noise module!
This is it, on the road to some nice functions for determining the
significant detection of a radar! Instead of relying on simple
thresholds (e.g., NCP < 0.3), we make use of the fact that the Doppler
velocity and spectrum width fields are available everywhere, and that
their texture fields (coherency) are a good at separating noise and
good signal, especially in low SNR and low NCP. However there are
certain  artifacts we have to be mindful of (e.g., velocity aliasing).
As it stands now this routine for determing significant detection
through velocity and spectrum width coherency is fully automatic. More
testing needs to be done, but I'm optimistic!