WIP: moment fields and processing
Moment fields and their processing now include my recent work on using
the Doppler velocity and spectrum width to determine the significant
detection of the radar, which avoids using an NCP threshold altogether.
I have also included another way to process the moment fields which
drastically speeds up processing time. This now needs to be done to the
texture fields and their processing!