WIP: update texture fields
Texture fields and its associated processing now can use the work I've
done on significant detection with the coherency fields. An update to
the textures histogram figure is thus in order. Finally, minimum and
maximum ranges can be specified when computing texture fields (e.g., to
avoid ground clutter close to the radar). Some minor changes are likely
to follow...