WIP: noise module
Major additions including the Hildebrand and Sekhon (1974) algorithm to
estimate the noise floor, as well as some other functions for filling
holes and/or dilation of binary masks. Furthermore, I am making some
changes to the data types of both texture fields (np.float64 ->
np.float32) and mask fields (np.bool -> np.int8). This is primarily to
make sure that we can successfully write NetCDF files (e.g., np.bool is
not a supported data type) as well as to decrease the size of the
written NetCDF files.