WIP: noise from velocity coherency
Big change in my velocity coherency algorithm which attempts to
characterize the significant detection of the radar by using the Doppler
velocity texture field. This algorithm now uses the predicted result
from Gaussian noise theory to determine which peak in the velocity
texture distribution corresponds to the noise peak. Previously I was
relying on the assumption that the noise peak was always at a higher
velocity texture than the coherent part of the velocity texture
distribution. However I came across examples where, if velocity aliasing
had a very strong signal, than this was no longer the case and results
were poor.